rename democracy all votes to flattened votes
