+++ packages/kintsugi-next/components/referenda/vote.js
@@ -38,11 +38,11 @@ const Popup = dynamic(() => import("./popup"), {
   ssr: false,
 });
 
-const AllVotesPopup = dynamic(
-  () => import("next-common/components/democracy/allVotesPopup"),
+const FlattenedVotesPopup = dynamic(
+  () => import("next-common/components/democracy/flattenedVotesPopup"),
   {
     ssr: false,
-  }
+  },
 );
 
 const Wrapper = styled.div`
@@ -362,7 +362,7 @@ function Vote({
       )}
 
       {showVoteList && (
-        <AllVotesPopup
+        <FlattenedVotesPopup
           setShowVoteList={setShowVoteList}
           allAye={allAye}
           allNay={allNay}
