+++ packages/next-common/components/democracy/flattenedVotesPopup/voteLabel.js
