+++ packages/next-common/components/summary/democracySummaryDelegation/democracySummaryDelegationInfo.js
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { useChainSettings } from "next-common/context/chain";
 import { toPrecision } from "next-common/utils";
-import VoteLabel from "next-common/components/democracy/allVotesPopup/voteLabel";
+import VoteLabel from "next-common/components/democracy/flattenedVotesPopup/voteLabel";
 import ValueDisplay from "next-common/components/valueDisplay";
 import User from "next-common/components/user";
 import { Conviction, ConvictionSupport } from "../../../utils/referendumCommon";
